workshop 2 task
We are going to make a simple responsive list of Instagram images and a
list of who has liked these images by using the  Instagram REST API. We
are going to use Bootstrap as responsive framework. The webapp we set up
should be scaffolded using  Yeoman and the webapp generator.

Note! The node_modules had to long filepaths, so it had to be zipped.
Unzip this if you're trying to run this